+++ CHANGELOG.rst
@@ -14,6 +14,17 @@ Change Log
 Unreleased
 **********
 
+[3.6.3] - 2023-01-27
+********************
+Fixed
+=======
+* In the consumer loop, open new database connection if the old one has closed before processing messages
+
+[3.6.2] - 2023-01-26
+********************
+Fixed
+=====
+* Reset mode now commits the correct offsets to Kafka.
 
 [3.6.1] - 2023-01-20
 ********************